requirements.txt
google-cloud-aiplatform
google-cloud-storage
firebase-admin
google-cloud-firestore
google-cloud-speech
vertexai

pandas
pyarrow
requests
requests-toolbelt

sentence-transformers
langchain
langchain-community

librosa
soundfile
scipy
pydub

pillow

Flask

